Update to  "@ohif/extension-dicom-ecg": "^1.0.4"
