+++ extensions/dicom-ecg/src/constants/graphicCalibration.js
@@ -1,0 +1,22 @@
+const uV = {
+  min: -500,
+  max: 1500,
+  deltaMain: 500,
+  deltaSecondary: 100,
+};
+
+const mV = {
+  min: -0.5,
+  max: 1.5,
+  deltaMain: 0.5,
+  deltaSecondary: 0.1,
+};
+
+const mmHg = {
+  min: 0,
+  max: 200,
+  deltaMain: 100,
+  deltaSecondary: 20,
+};
+
+export default { uV, mV, mmHg };
