+++ platform/viewer/package.json
@@ -53,7 +53,7 @@
     "@ohif/extension-dicom-html": "^1.3.22",
     "@ohif/extension-dicom-microscopy": "^0.52.0",
     "@ohif/extension-dicom-pdf": "^1.1.2",
-    "@ohif/extension-dicom-ecg": "^1.0.3",
+    "@ohif/extension-dicom-ecg": "^1.0.4",
     "@ohif/extension-dicom-rt": "^0.7.11",
     "@ohif/extension-dicom-segmentation": "^0.7.11",
     "@ohif/extension-lesion-tracker": "^0.2.2",
